move test into src and update build configs (#3)
move test into src and update build configs
