+++ jest.config.js
@@ -4,10 +4,8 @@ module.exports = {
       tsConfig: 'tsconfig.json',
     },
   },
-  moduleFileExtensions: ['ts', 'js'],
   transform: {
     '^.+\\.(ts|tsx)$': 'ts-jest',
   },
-  testMatch: ['**/test/**/*.test.(ts|js)'],
   testEnvironment: 'node',
 };
